requirements.txt
amqp==2.1.4
billiard==3.5.0.2
bunch==1.0.1
celery==4.0.2
certifi==2016.9.26
cffi==1.9.1
cryptography==1.5.3
dj-database-url==0.4.1
Django==1.10.3
django-braces==1.10.0
django-mysql==1.1.0
django-nose==1.4.4
djangorestframework==3.4.6
EasyProcess==0.2.3
enum34==1.1.6
funcsigs==1.0.2
google-api-python-client==1.5.1
gspread==0.4.1
gunicorn==19.6.0
httplib2==0.9.2
idna==2.1
ipaddress==1.0.17
kombu==4.0.2
mock==2.0.0
MySQL-python==1.2.5
nose==1.3.7
numpy==1.11.2
oauth2client==3.0.0
pandas==0.19.2
pbr==1.10.0
Pillow==3.4.2
poster==0.8.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pyOpenSSL==16.2.0
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.11
redis==2.10.5
requests==2.11.1
rsa==3.4.2
selenium==3.0.1
simplejson==3.10.0
six==1.10.0
SQLAlchemy==1.1.4
tldextract==1.6
transitions==0.4.3
Unipath==1.1
uritemplate==0.6
urllib3==1.19.1
vine==1.1.3
whitenoise==3.2.2
